[Security Solution] Migrate Entities Details to Flyout v2 and OneDiscover
Moves host and user entity details out of the legacy expandable flyout's
context-coupled left panel and into a standalone Flyout v2 tool surface,
enabling the same content to render in the Security flyout, Flyout v2,
and Discover via the OneDiscover bridge.
