+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/tools/entities/test_ids.ts
@@ -1,0 +1,13 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { PREFIX } from '../../../../flyout/shared/test_ids';
+
+export const ENTITIES_DETAILS_TEST_ID = `${PREFIX}EntitiesDetails` as const;
+export const USER_DETAILS_TEST_ID = `${PREFIX}UsersDetails` as const;
+export const HOST_DETAILS_TEST_ID = `${PREFIX}HostsDetails` as const;
+export const ENTITY_ALERTS_DETAILS_TEST_ID = `${PREFIX}EntityAlertsDetails` as const;
